hw4

+++ controllers/contactsController.js
@@ -4,7 +4,11 @@ import { HttpError } from '../helpers/HttpError.js'
 
 const getAll = async (req, res, next) => {
     try {
-        const data = await Contact.find()
+        const { _id: owner } = req.user;
+        const { page = 1, limit = 10, ...filterParams } = req.query;
+        const skip = (page - 1) * limit;
+        const filter = { owner, ...filterParams };
+        const data = await Contact.find(filter, "-createdAt -updatedAt", { skip, limit }).populate("owner", "username email");
         res.json(data)
     } catch (error) {
         next(error)
@@ -13,27 +17,23 @@ const getAll = async (req, res, next) => {
 
 const getById = async (req, res, next) => {
     try {
+        const { _id: owner } = req.user;
         const { contactId } = req.params
-        const data = await Contact.findById(contactId)
+        const data = await Contact.findById({ _id: contactId, owner })
         if (!data) {
             throw HttpError(404, 'Contact not found');
         }
         res.json(data)
     }
     catch (error) {
-        console.log(error)
         next(error)
     }
 }
 
 const addContact = async (req, res, next) => {
     try {
-        const contact = req.body
-        const { error } = addContactSchema.validate(contact)
-        if (error) {
-            throw HttpError(400, error.message)
-        }
-        const data = await Contact.create(contact)
+        const { _id: owner } = req.user;
+        const data = await Contact.create({ ...req.body, owner })
         res.status(201).json(data)
     }
     catch (error) {
@@ -43,8 +43,9 @@ const addContact = async (req, res, next) => {
 
 const removeContact = async (req, res, next) => {
     try {
+        const { _id: owner } = req.user;
         const { contactId } = req.params
-        const data = await Contact.findByIdAndDelete(contactId)
+        const data = await Contact.findByIdAndDelete({ _id: contactId, owner })
         if (!data) {
             throw HttpError(404, 'Contact not found');
         }
@@ -56,8 +57,9 @@ const removeContact = async (req, res, next) => {
 
 const updateContact = async (req, res, next) => {
     try {
+        const { _id: owner } = req.user;
         const { contactId } = req.params
-        const data = await Contact.findByIdAndUpdate(contactId, req.body)
+        const data = await Contact.findByIdAndUpdate({ _id: contactId, owner }, req.body)
         if (!data) {
             throw HttpError(404, 'Contact not found');
         }
